+++ app.R
@@ -7,16 +7,8 @@
 # renv::install("janlisec/fioRa")
 
 # To deploy to shinyapps.io:
-# rsconnect::deployApp(appDir = "C:/Users/jlisec/Documents/Rpackages/Rpackage_fioRa/fioRa", appName = "fioraApp", forceUpdate = TRUE)
-
-# To check for errors in shinyapps.io:
-# rsconnect::showLogs(appName = "fioraApp", account = "jali")
-
+# This app should not be deployed to shinyapps.io as it requires a python conda environment which takes very long to install using reticulate on shinyapps.io
 
 pkgload::load_all(export_all = FALSE, helpers = FALSE, attach_testthat = FALSE)
 options("golem.app.prod" = TRUE)
-# this should be TRUE only for deployment to shinyapps.io
-#options(fiora.deploy_to_shinyapps = tryCatch(rsconnect::serverInfo(name = "shinyapps.io")$name, error = function(e) "")=="shinyapps.io")
-#options(fiora.deploy_to_shinyapps = TRUE)
-options(fiora.deploy_to_shinyapps = FALSE)
 run_app() # add parameters here (if any)