+++ NEWS.md
@@ -1,3 +1,13 @@
+# v0.2.4 by jan.lisec@bam.de
+
+* predict() renamed to run_script()
+* detection of python and script path platform specific
+* script evaluation platform specific
+
+# v0.2.3 by jan.lisec@bam.de
+
+* script path hardcoded; reticulate not used
+
 # v0.2.2 by jan.lisec@bam.de
 
 * moved reticulate to Suggest in Description if python installation and script is found